pyproject.toml
[project]
name = "agentsmicroservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.129.0",
    "github-copilot-sdk>=0.1.25",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
]
